Create async board generation logic, store game creattion logic in separate helper file
